requirements.txt
Flask==2.3.2
flask_cors==4.0.0
matplotlib==3.7.1
numpy==1.23.5
numpy==1.24.4
object_detection==0.0.3
opencv_python==4.7.0.72
opencv_python_headless==4.8.1.78
Pillow==9.5.0
Pillow==10.1.0
protobuf==4.25.1
pymongo==4.6.1
pytesseract==0.3.10
Requests==2.31.0
scipy==1.11.4
six==1.16.0
skimage==0.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
